Moved cilMaskPolyData into visualisation_pipeline
